+++ wagtail/admin/forms/__init__.py
@@ -2,7 +2,6 @@ import copy
 from itertools import groupby
 
 from django import forms
-from django.contrib.auth.forms import AuthenticationForm, PasswordResetForm
 from django.contrib.auth.models import Group, Permission
 from django.core import validators
 from django.db import models, transaction
@@ -18,6 +17,8 @@ from wagtail.admin import widgets
 from wagtail.core.models import (
     BaseViewRestriction, Collection, CollectionViewRestriction, GroupCollectionPermission, Page,
     PageViewRestriction)
+
+from .auth import *  # NOQA
 
 
 class URLOrAbsolutePathValidator(validators.URLValidator):
@@ -59,34 +60,6 @@ class ExternalLinkChooserForm(forms.Form):
 class EmailLinkChooserForm(forms.Form):
     email_address = forms.EmailField(required=True)
     link_text = forms.CharField(required=False)
-
-
-class LoginForm(AuthenticationForm):
-    username = forms.CharField(
-        max_length=254, widget=forms.TextInput(attrs={'tabindex': '1'}))
-
-    password = forms.CharField(
-        widget=forms.PasswordInput(attrs={
-            'tabindex': '2',
-            'placeholder': ugettext_lazy("Enter password"),
-        }))
-
-    def __init__(self, request=None, *args, **kwargs):
-        super().__init__(request=request, *args, **kwargs)
-        self.fields['username'].widget.attrs['placeholder'] = (
-            ugettext_lazy("Enter your %s") % self.username_field.verbose_name)
-
-    @property
-    def extra_fields(self):
-        for field_name, field in self.fields.items():
-            if field_name not in ['username', 'password']:
-                yield field_name, field
-
-
-class PasswordResetForm(PasswordResetForm):
-    email = forms.EmailField(
-        label=ugettext_lazy("Enter your email address to reset your password"),
-        max_length=254, required=True)
 
 
 class CopyForm(forms.Form):
